+++ poko-compiler-plugin/src/main/kotlin/dev/drewhamilton/poko/fir/PokoFirDeclarationGenerationExtension.kt
@@ -8,6 +8,7 @@ import org.jetbrains.kotlin.descriptors.ClassKind
 import org.jetbrains.kotlin.descriptors.Modality
 import org.jetbrains.kotlin.descriptors.Visibilities
 import org.jetbrains.kotlin.fir.FirSession
+import org.jetbrains.kotlin.fir.declarations.FirNamedFunction
 import org.jetbrains.kotlin.fir.declarations.processAllDeclaredCallables
 import org.jetbrains.kotlin.fir.declarations.utils.isExtension
 import org.jetbrains.kotlin.fir.declarations.utils.isFinal
@@ -17,6 +18,7 @@ import org.jetbrains.kotlin.fir.extensions.FirDeclarationPredicateRegistrar
 import org.jetbrains.kotlin.fir.extensions.MemberGenerationContext
 import org.jetbrains.kotlin.fir.extensions.predicate.LookupPredicate
 import org.jetbrains.kotlin.fir.extensions.predicateBasedProvider
+import org.jetbrains.kotlin.fir.plugin.createMemberFunction
 import org.jetbrains.kotlin.fir.resolve.toClassSymbol
 import org.jetbrains.kotlin.fir.scopes.impl.FirClassDeclaredMemberScope
 import org.jetbrains.kotlin.fir.symbols.impl.FirClassSymbol
@@ -226,7 +228,7 @@ internal class PokoFirDeclarationGenerationExtension(
 
     private fun createEqualsFunction(
         owner: FirClassSymbol<*>,
-    ): FirSimpleFunctionCompat = createMemberFunctionCompat(
+    ): FirNamedFunction = createMemberFunction(
         owner = owner,
         key = PokoKey,
         name = Equals.functionName,
@@ -245,7 +247,7 @@ internal class PokoFirDeclarationGenerationExtension(
 
     private fun createHashCodeFunction(
         owner: FirClassSymbol<*>,
-    ): FirSimpleFunctionCompat = createMemberFunctionCompat(
+    ): FirNamedFunction = createMemberFunction(
         owner = owner,
         key = PokoKey,
         name = HashCode.functionName,
@@ -256,7 +258,7 @@ internal class PokoFirDeclarationGenerationExtension(
 
     private fun createToStringFunction(
         owner: FirClassSymbol<*>,
-    ): FirSimpleFunctionCompat = createMemberFunctionCompat(
+    ): FirNamedFunction = createMemberFunction(
         owner = owner,
         key = PokoKey,
         name = ToString.functionName,